pyproject.toml
[project]
name = "urban-hire"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.1.2",
    "django>=6.0.4",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "python-dotenv>=1.1.0",
    "psycopg>=3.3.3",
    "psycopg2-binary>=2.9.11",
]
